requirements.txt
openai>=1.0.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
xgboost>=1.6.0
joblib>=1.1.0
python-dateutil>=2.8.0
streamlit>=1.25.0
python-dotenv>=1.0.0
